Settings: Add an option to force pre-O apps to use full screen aspect ratio
When an app target pre-O releases, the default max aspect ratio
is 1.86:1 which leads to ugly black areas on devices that have
screens with higher aspect ratio (for example Galaxy S8/S9).

This change adds an option to allow users to change aspect ratio
for pre-O apps to full screen aspect ratio.

    From: Erfan Abdi <erfangplus@gmail.com>
    Date: Wed, 9 Mar 2022 23:04:46 +0330
    Subject: [PATCH] Implement LongScreen settings

    Based on https://github.com/LineageOS/android_packages_apps_LineageParts

    Change-Id: I17670779dabaf93ce9d9b2525db269cfea02c35e
    Signed-off-by: Mohammad Hasan Keramat J <ikeramat@protonmail.com>

Change-Id: I0fb73511cf654ee22a4cfb7aef252008f8db8855
Co-authored-by: Erfan Abdi <erfangplus@gmail.com>
Signed-off-by: Mohammad Hasan Keramat J <ikeramat@protonmail.com>